requirements.txt
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.1
websockets>=12.0
playwright>=1.40.0
